feat: add instructor online class components
